refactored based of a datasets module
